+++ v2/src/core/components/AppInput.vue
@@ -17,12 +17,12 @@
                 clearable
                 @update:modelValue="onUserInput"
                 @click:clear="resetValues"
-                bg-color="greyCard"
+                bg-color="lightGrey"
                 color="primary"
             >
                 <template v-if="hasPreppendInner" #prepend-inner>
                     <slot name="prepend">
-                        <v-icon :color="state.value ? (props.hasError ? 'error' : 'secondary') : 'greyInputText'" icon="search" />
+                        <v-icon :color="state.value ? (props.hasError ? 'error' : 'secondary') : 'info'" icon="search" />
                     </slot>
                 </template>
                 <template v-else #prepend-inner>