+++ v2/src/views/ViewTokens.vue
@@ -1,6 +1,9 @@
 <template>
     <v-row :class="rowMargin">
         <v-col cols="12" :class="columnPadding">
+            <module-big-movers />
+        </v-col>
+        <v-col cols="12" :class="columnPadding">
             <app-ad-buttons-large />
         </v-col>
         <v-col cols="12" :class="columnPadding">
@@ -11,12 +14,18 @@
 
 <script setup lang="ts">
 import AppAdButtonsLarge from '@/core/components/AppAdButtonsLarge.vue'
+import ModuleBigMovers from '@/modules/tokens/ModuleBigMovers.vue'
 import ModuleTokensInfo from '@module/tokens/ModuleTokensInfo.vue'
 import { useAppViewGrid } from '@core/composables/AppViewGrid/AppViewGrid.composable'
 import { onMounted } from 'vue'
 import { onBeforeRouteUpdate } from 'vue-router'
 import { tabViewRouteGuardOnUpdate } from '@/core/router/helpers'
 import { Q_TOKENS } from '@/core/router/routesNames'
+import { VIEW_TAGS } from '@core/helper/tags'
+import { usePageMeta } from '@core/composables/use-page-meta/use-page-meta.composable'
+
+usePageMeta(null, VIEW_TAGS.TOKENS)
+
 onMounted(() => {
     window.scrollTo(0, 0)
 })